add 'spacer' boxes
Certain sounds (like 'yi' or 'ye') aren't present in the Japanese
language, so hiragana/katakana charts commonly just leave a space where
they would be.
I did the same, but some people not knowing this thought there was a
bug causing these sounds to not display.
I added spacer boxes to show that this was in fact intended. (: